move stores into sveltik context
The stores were shared globally which means only one form could work on the page at one time.
Moving the stores into the context allows you to have multiple sveltik contexts i.e. forms
on the page at the same time.

A side effect of this is that Field only works when inside a Sveltik context, hence the changes
to the tests.